requirements.txt
fastapi==0.108.0
httpx==0.26.0
passlib[bcrypt]==1.7.4
pytest==7.4.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
SQLAlchemy==2.0.24
uvicorn[standard]==0.25.0
